Fix issue #17: bytes-to-array makes a copy of byte buffer
The `ucar.ma2.Array/factory` method sometimes peeks at the underlying
`byte[]` storage of a `ByteBuffer` argument instead of relying on the
`ByteBuffer` interface. As a result, the `Array` can be constructed
using bytes that are not even exposed by the `ByteBuffer`.

This commit changes the implementation of the `bytes-to-array` function
in the `io.mandoline.impl` namespace to be robust against this bug. If
the byte buffer is not perfectly aligned with its underlying byte array,
then a copy is created. The copy is backed by a byte array that only
contains the bytes that the original byte buffer exposes, starting from
its current position and ending at its limit.